Collapse skipper
It is really internal thing of BF skipper, no need for
iface + 2 separate classes.

Expose and adjust tests.
